test: migrate config/api tests to vitest
Migrates the API tests to vitest.

Snapshot has been verified using `npx ava-vitest-snapshot-diff {old} {new}`
